backend/requirements.txt
flask
flask-cors
mysql-connector-python
pyjwt
google-genai
Pillow
passlib